Allow 8 octets input signed extension function (64bit) (#151)
* Allow 8 octets input signed extension function (64bit)

Allow up to 8 input octets for the signed extension function as a consequence of the change to a 64bit architecture. Direct evidence of this requirement in instructions 140 & 141.

* GP-0.5.0-eq:A.17 (32bit -> 64bit)

Allow `L_Y` to be up to (value) 8 and thus allow the `signed extension`-function up to 8 bytes of input to determine `V_Y`.

* GP-0.5.0-eq:A.20 (32bit -> 64bit)

In GP-0.5.0-eq:A.20: Allow `L_Y` to be up to (value) 8 and thus allow the `signed extension`-function up to 8 bytes of input to determine `V_Y`. And add mod2^32 to instruction 72.

* GP-0.5.0-eq:A.21 (32bit -> 64bit)

In GP-0.5.0-eq:A.21: Allow `L_X` to be up to (value) 8 and thus allow the `signed extension`-function up to 8 bytes of input to determine `V_X`.

* GP-0.5.0-eq:A.23 (32bit -> 64bit)

In GP-0.5.0-eq:A.23: Allow `L_X` to be up to (value) 8 and thus allow the `signed extension`-function up to 8 bytes of input to determine `V_X`. Required for instructions: 126, 127, 132, 133, 137, 138, 139, 140, 141, 142, 144, 145, 146, 147.

* Corrections pvm.tex

Corrections